+++ Sources/SmileID/Classes/SmileID.swift
@@ -3,7 +3,7 @@ import SwiftUI
 import UIKit
 
 public class SmileID {
-    public static let version = "10.0.0-beta10"
+    public static let version = "10.0.0-beta11"
     @Injected var injectedApi: SmileIDServiceable
     public static var configuration: Config { config }
 
@@ -114,6 +114,7 @@ public class SmileID {
     ///   - showAttribution: Whether to show the Smile ID attribution or not on the Instructions
     ///     screen
     ///   - showInstructions: Whether to deactivate capture screen's instructions for SmartSelfie.
+    ///   - partnerParams: Custom values specific to partners
     ///   - delegate: Callback to be invoked when the SmartSelfie™ Enrollment is complete.
     public class func smartSelfieEnrollmentScreen(
         userId: String = generateUserId(),
@@ -121,12 +122,14 @@ public class SmileID {
         allowAgentMode: Bool = false,
         showAttribution: Bool = true,
         showInstructions: Bool = true,
+        partnerParams: [String: String] = [:],
         delegate: SmartSelfieResultDelegate
     ) -> some View {
         let viewModel = SelfieCaptureViewModel(
             userId: userId,
             jobId: jobId,
             isEnroll: true,
+            partnerParams: partnerParams,
             allowsAgentMode: allowAgentMode,
             showAttribution: showAttribution
         )
@@ -153,6 +156,7 @@ public class SmileID {
     ///   - showAttribution: Whether to show the Smile ID attribution or not on the Instructions
     ///     screen
     ///   - showInstructions: Whether to deactivate capture screen's instructions for SmartSelfie.
+    ///   - partnerParams: Custom values specific to partners
     ///   - delegate: Callback to be invoked when the SmartSelfie™ Authentication is complete.
     public class func smartSelfieAuthenticationScreen(
         userId: String,
@@ -160,12 +164,14 @@ public class SmileID {
         allowAgentMode: Bool = false,
         showAttribution: Bool = true,
         showInstructions: Bool = true,
+        partnerParams: [String: String] = [:],
         delegate: SmartSelfieResultDelegate
     ) -> some View {
         let viewModel = SelfieCaptureViewModel(
             userId: userId,
             jobId: jobId,
             isEnroll: false,
+            partnerParams: partnerParams,
             allowsAgentMode: allowAgentMode,
             showAttribution: showAttribution
         )
@@ -196,6 +202,7 @@ public class SmileID {
     ///   - showInstructions: Whether to deactivate capture screen's instructions for Document
     ///   Verification (NB! If instructions are disabled, gallery upload won't be possible)
     ///   - showAttribution: Whether to show the Smile ID attribution on the Instructions screen
+    ///   - partnerParams: Custom values specific to partners
     ///   - delegate: The delegate object that receives the result of the Document Verification
     public class func documentVerificationScreen(
         userId: String = generateUserId(),
@@ -209,6 +216,7 @@ public class SmileID {
         allowGalleryUpload: Bool = false,
         showInstructions: Bool = true,
         showAttribution: Bool = true,
+        partnerParams: [String: String] = [:],
         delegate: DocumentVerificationResultDelegate
     ) -> some View {
         OrchestratedDocumentVerificationScreen(
@@ -223,6 +231,7 @@ public class SmileID {
             allowGalleryUpload: allowGalleryUpload,
             allowAgentMode: allowAgentMode,
             showInstructions: showInstructions,
+            partnerParams: partnerParams,
             onResult: delegate
         ).environmentObject(router)
     }
@@ -251,6 +260,7 @@ public class SmileID {
     ///   - showInstructions: Whether to deactivate capture screen's instructions for Document
     ///   Verification (NB! If instructions are disabled, gallery upload won't be possible)
     ///   - showAttribution: Whether to show the Smile ID attribution on the Instructions screen
+    ///   - partnerParams: Custom values specific to partners
     ///   - delegate: The delegate object that receives the result of the Document Verification
     public class func enhancedDocumentVerificationScreen(
         userId: String = generateUserId(),
@@ -264,6 +274,7 @@ public class SmileID {
         allowGalleryUpload: Bool = false,
         showInstructions: Bool = true,
         showAttribution: Bool = true,
+        partnerParams: [String: String] = [:],
         delegate: EnhancedDocumentVerificationResultDelegate
     ) -> some View {
         OrchestratedEnhancedDocumentVerificationScreen(
@@ -278,6 +289,7 @@ public class SmileID {
             allowGalleryUpload: allowGalleryUpload,
             allowAgentMode: allowAgentMode,
             showInstructions: showInstructions,
+            partnerParams: partnerParams,
             onResult: delegate
         ).environmentObject(router)
     }
@@ -302,6 +314,7 @@ public class SmileID {
     ///   the front camera will be used.
     ///  - showAttribution: Whether to show the Smile ID attribution on the Instructions screen
     ///  - showInstructions: Whether to deactivate capture screen's instructions for SmartSelfie.
+    ///  - partnerParams: Custom values specific to partners
     ///  - delegate: Callback to be invoked when the Biometric KYC is complete.
     public class func biometricKycScreen(
         partnerIcon: UIImage,
@@ -314,6 +327,7 @@ public class SmileID {
         allowAgentMode: Bool = false,
         showAttribution: Bool = true,
         showInstructions: Bool = true,
+        partnerParams: [String: String] = [:],
         delegate: BiometricKycResultDelegate
     ) -> some View {
         OrchestratedBiometricKycScreen(
@@ -327,6 +341,7 @@ public class SmileID {
             showInstructions: showInstructions,
             showAttribution: showAttribution,
             allowAgentMode: allowAgentMode,
+            partnerParams: partnerParams,
             delegate: delegate
         )
     }