+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - BVN OTP API calls and models
+- Default to `production` on `SmileID.initialize()`
 
 ### Changed