+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Consent Screen
 - BVN OTP API calls and models
+- Default to `production` on `SmileID.initialize()`
 
 ### Changed